# docs/swagger.yaml
definitions:
...
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_searchhistory.SearchHistory:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /history:
    get:
      consumes:
      - application/json
      description: Get all search history for a user by ID
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_searchhistory.SearchHistory'
            type: array
        "401":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get Search History
      tags:
      - search-history
    post:
      consumes:
      - application/json
      description: Add a new entry to the search history for a user
      parameters:
      - description: Name of the search history entry
        in: body
        name: name
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Add Search History
      tags:
      - search-history
...
swagger: "2.0"